Multiple changes.
 * Show line numbers in file content view.
 * Fix factory tests no to use the `Text` type.
 * Add argument validation for the sleep function.
 * Bump version to 3.4.4.6
 * VfsOpen: fix handling of `implementation_type` in combination with `GetRawDevice`.
 * Work around Unicode issues in the cloud metadata action.
 * action_test: fix test failure.
 * Improve immutability guarantees of tests.
 * Replace anonymous Store providers with centralized STORE_PROVIDERS.
 * Centralize MockStore providers.
 * Replace brittle ActivatedRoute mocks with real routing.
 * Link to Drawer from file results.
 * Implement [drawerLink] directive to open drawer.
 * Server side flows: propagate `implementation_type`.
 * MSI: remove legacy nanny service (if it exists) in fleetspeak {enabled,bundled} configuration.
 * Remove deprecated datastore config options.
 * Bump version to 3.4.4.5
 * VfsFileFinder: more propagation of `implementation_type`.
 * Test UIv2 file content viewing with Selenium.
 * Implement basic file view.
 * Implement enum preconditions.
 * client_utils_osx: fix endless loop in `GetRawDevice`.
 * Enable apfs for TSK on macOS.
 * Update pytsk3 dep to 20210419 (same version as in third_party) on macOS.
 * Migrate ngOnDestroy to observeOnDestroy.
 * Implement observeOnDestroy helper to unsubscribe from Subscriptions in ngOnDestroy.
 * Bump version to 3.4.4.4
 * Sandboxing: send memory chunks to unprivileged process in batches.
 * Fix open-source test failure for undefined ActivatedRouteSnapshot.children.
 * Rename mock_store to store_test_util.
 * Create Store, API method, and Component stub for file content viewing.
 * Replace custom Store mocks with MockStore.
 * Introduce MockStore to reduce boilerplate of Store mocks.
 * Sandboxing: disable memory sandboxing on OSX.
 * Sandboxing: correctly propagate `PathSpec.implementation_type`.
 * Improve robustness of the timeline end-to-end tests.
 * Improve test robustness by replacing Object.defineProperty.
 * Centralize ClientID selection in SelectedClientGlobalStore.
 * Extract MatDrawer to App-wide, URL-based auxiliary socket.
 * Sandboxing: add security mechanism tests for Linux and OSX.
 * Remove __iadd__ footgun.
 * Compare equality only for RDFStruct fields with type descriptors.
 * client_vfs_test.py: Flush files cache to avoid permission errors on Windows.
 * Mitigate AttributeErrors in equality comparison of RDFStructs of different versions.
 * Sandboxing: add test for security mechanisms on Windows.
 * Ignore StatEntry results in ListProcess details.
 * Implement checkbox for collect_binaries argument.
 * Updating psutil version in setup.py to 5.8.0.
 * Implement NetworkConnection.State filter for ListProcesses flow.

+++ grr/client/grr_response_client/unprivileged/windows/process.py
@@ -45,6 +45,7 @@ import subprocess
 from typing import List, Optional
 
 import win32api
+import win32con
 import win32event
 import win32process
 
@@ -302,9 +303,29 @@ class Process:
       if self.pid == 0:
         raise Error("GetProcessId failed.")
 
-  def Stop(self) -> None:
-    win32process.TerminateProcess(self._handle, -1)
+  def Stop(self) -> int:
+    """Terminates the process and waits for the process to exit.
+
+    Returns:
+      The exit code.
+    """
+    exit_code = win32process.GetExitCodeProcess(self._handle)
+    if exit_code == win32con.STILL_ACTIVE:
+      win32process.TerminateProcess(self._handle, -1)
+    return self.Wait()
+
+  def Wait(self) -> int:
+    """Waits for the process to exit.
+
+    Returns:
+      The exit code.
+
+    Raises:
+      Error: on system error.
+    """
     res = win32event.WaitForSingleObject(self._handle, win32event.INFINITE)
     if res == win32event.WAIT_FAILED:
       raise Error("WaitForSingleObject failed.")
+    exit_code = win32process.GetExitCodeProcess(self._handle)
     self._exit_stack.close()
+    return exit_code
